+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Fixed a bug where the notebook check raised ModuleNotFoundError when IPython not installed
 - Fixed a memory leak with metrics that causes issues with very long epochs
 - Fixed a bug with the once and once_per_epoch decorators
+- Fixed a bug where the test criterion wouldn't accept a function of state
 - Fixed a bug where type inference would not work correctly when chaining ``Trial`` methods
 
 ## [0.3.0] - 2019-02-28